requirements.txt
cffi==1.17.1
prettyprint==0.1.5
pycparser==2.22
raylib==5.5.0.2
tomli_w==1.1.0
